requirements.txt
aiomysql==0.2.0
python-dotenv==1.0.1
pydantic[email,timezone]==2.10.3
pydantic-settings==2.7.0
SQLAlchemy==2.0.36
fastapi==0.115.6
uvicorn[standard]==0.33.0
black==24.10.0
isort==5.13.2
alembic==1.14.0
greenlet==3.1.1
boto3==1.35.81
asyncpg==0.30.0
python-keycloak==5.1.1
python-multipart==0.0.20
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.4.0
aiohttp==3.11.11
Jinja2==3.1.5
redis==5.2.1
httpx==0.28.1
pytest-asyncio==0.25.3
celery[redis]==5.4.0
sqlalchemy-celery-beat==0.8.0
asgiref==3.8.1
ruff==0.11.2
pytest-mock==3.14.0
httpx==0.28.1
aio-pika==9.5.5
loguru==0.7.3
babel==2.17.0
python-slugify==8.0.4
GeoAlchemy2==0.17.1